# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Cyvadra_toktik_internal_dto.BrowserColumn:
    properties:
      codec_expression:
        type: string
      comment:
        type: string
      default_expression:
        type: string
      default_kind:
        type: string
      is_nullable:
        type: boolean
      name:
        type: string
      position:
        type: integer
      type:
        type: string
    type: object
  github_com_Cyvadra_toktik_internal_dto.BrowserCoverageResponse:
    properties:
      daily:
        items:
          $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.BrowserDailyCoverage'
        type: array
      dataset:
        $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.BrowserDatasetDescriptor'
      first_timestamp:
        type: string
      last_timestamp:
        type: string
      row_count:
        type: integer
    type: object
  github_com_Cyvadra_toktik_internal_dto.BrowserDailyCoverage:
    properties:
      date:
        type: string
      row_count:
        type: integer
    type: object
  github_com_Cyvadra_toktik_internal_dto.BrowserDatasetDescriptor:
    properties:
      checks:
        items:
          type: string
        type: array
      fields:
        items:
          type: string
        type: array
      market:
        type: string
      name:
        type: string
      relation:
        type: string
      symbol_field:
        type: string
      time_field:
        type: string
    type: object
  github_com_Cyvadra_toktik_internal_dto.BrowserFieldProfileResponse:
    properties:
      dataset:
        $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.BrowserDatasetDescriptor'
      distinct_count:
        type: integer
      empty_count:
        type: integer
      field:
        type: string
      max: {}
      min: {}
      null_count:
        type: integer
      row_count:
        type: integer
      type:
        type: string
      zero_count:
        type: integer
    type: object
  github_com_Cyvadra_toktik_internal_dto.BrowserPresetResponse:
    properties:
      datasets:
        items:
          $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.BrowserDatasetDescriptor'
        type: array
    type: object
  github_com_Cyvadra_toktik_internal_dto.BrowserPreviewResponse:
    properties:
      columns:
        items:
          type: string
        type: array
      data:
        items:
          additionalProperties: {}
          type: object
        type: array
      dataset:
        $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.BrowserDatasetDescriptor'
    type: object
  github_com_Cyvadra_toktik_internal_dto.BrowserSchemaResponse:
    properties:
      columns:
        items:
          $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.BrowserColumn'
        type: array
      dataset:
        $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.BrowserDatasetDescriptor'
    type: object
  github_com_Cyvadra_toktik_internal_dto.BrowserValidCountResponse:
    properties:
      check:
        type: string
      dataset:
        $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.BrowserDatasetDescriptor'
      invalid_count:
        type: integer
      row_count:
        type: integer
      valid_count:
        type: integer
    type: object
...
  github_com_Cyvadra_toktik_internal_dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:9010
info:
  contact:
    email: dev@toktik.dev
    name: Toktik Dev Team
    url: https://toktik.dev
  description: Backend data services for multi-market analytics, including market
    data retrieval, feature/factor/fundamental queries, screening, strategy catalog,
    and backtesting.
  license:
    name: Proprietary
    url: https://toktik.dev/license
  termsOfService: https://toktik.dev/terms
  title: Toktik Options Platform API
  version: "1.0"
paths:
...
  /browser/datasets/{dataset}/coverage:
    get:
      description: Returns first/last timestamps and daily row counts for a server-approved
        browser dataset.
      parameters:
      - description: Dataset preset name
        in: path
        name: dataset
        required: true
        type: string
      - description: Symbol filter
        in: query
        name: symbol
        type: string
      - description: Underlying filter
        in: query
        name: underlying
        type: string
      - description: Start time
        in: query
        name: from
        type: string
      - description: End time
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.BrowserCoverageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.ErrorResponse'
      summary: Get dataset time coverage
      tags:
      - DataBrowser
  /browser/datasets/{dataset}/field-profile:
    get:
      description: Returns simple null/zero/empty/distinct/min/max stats for an approved
        field.
      parameters:
      - description: Dataset preset name
        in: path
        name: dataset
        required: true
        type: string
      - description: Field name
        in: query
        name: field
        required: true
        type: string
      - description: Start time
        in: query
        name: from
        type: string
      - description: End time
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.BrowserFieldProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.ErrorResponse'
      summary: Profile a dataset field
      tags:
      - DataBrowser
  /browser/datasets/{dataset}/preview:
    get:
      description: Returns a bounded row sample for a server-approved browser dataset.
      parameters:
      - description: Dataset preset name
        in: path
        name: dataset
        required: true
        type: string
      - description: Symbol filter
        in: query
        name: symbol
        type: string
      - description: Underlying filter
        in: query
        name: underlying
        type: string
      - description: Start time
        in: query
        name: from
        type: string
      - description: End time
        in: query
        name: to
        type: string
      - description: Comma-separated approved columns
        in: query
        name: columns
        type: string
      - description: Max rows, capped at 1000
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.BrowserPreviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.ErrorResponse'
      summary: Preview dataset rows
      tags:
      - DataBrowser
  /browser/datasets/{dataset}/schema:
    get:
      description: Returns ClickHouse column metadata for a server-approved browser
        dataset.
      parameters:
      - description: Dataset preset name
        in: path
        name: dataset
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.BrowserSchemaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.ErrorResponse'
      summary: Get dataset schema
      tags:
      - DataBrowser
  /browser/datasets/{dataset}/valid-count:
    get:
      description: Counts valid and invalid rows using a server-approved validity
        check.
      parameters:
      - description: Dataset preset name
        in: path
        name: dataset
        required: true
        type: string
      - description: Validity check name
        in: query
        name: check
        type: string
      - description: Start time
        in: query
        name: from
        type: string
      - description: End time
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.BrowserValidCountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.ErrorResponse'
      summary: Count valid rows
      tags:
      - DataBrowser
  /browser/presets:
    get:
      description: Returns server-approved datasets and checks available to the internal
        data browser.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.BrowserPresetResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Cyvadra_toktik_internal_dto.ErrorResponse'
      summary: List data browser presets
      tags:
      - DataBrowser
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"